requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
sentence-transformers
openai
python-dotenv
langsmith
tiktoken
streamlit